command to import courses from testr-1.0 and minor bug corrections

+++ backend/backend/views/submission_view.py
@@ -66,7 +66,8 @@ class AddSubmissionAPIView(APIView):
 
         # Procurar por submissão existente
         submission = Submission.objects.filter(question=question, student=student).first()
-        submission.delete()
+        if submission:
+            submission.delete()
 
         # Criar nova submissão
         serializer = SubmissionSerializer(data={
